requirements/common.txt
django-flat-theme==0.9.5
django-annoying==0.8.2
Pillow==2.9.0
